refactor: extract streaming sse transport
